+++ packages/engine/Source/Scene/Model/Model.js
@@ -1,23 +1,17 @@
-import AttributeType from "../AttributeType.js";
-import AttributeCompression from "../../Core/AttributeCompression.js";
 import BoundingSphere from "../../Core/BoundingSphere.js";
 import Cartesian3 from "../../Core/Cartesian3.js";
 import Cartographic from "../../Core/Cartographic.js";
 import Check from "../../Core/Check.js";
 import Credit from "../../Core/Credit.js";
 import Color from "../../Core/Color.js";
-import ComponentDatatype from "../../Core/ComponentDatatype.js";
 import defined from "../../Core/defined.js";
 import defaultValue from "../../Core/defaultValue.js";
 import DeveloperError from "../../Core/DeveloperError.js";
 import destroyObject from "../../Core/destroyObject.js";
 import DistanceDisplayCondition from "../../Core/DistanceDisplayCondition.js";
 import Event from "../../Core/Event.js";
-import IndexDatatype from "../../Core/IndexDatatype.js";
-import IntersectionTests from "../../Core/IntersectionTests.js";
 import Matrix3 from "../../Core/Matrix3.js";
 import Matrix4 from "../../Core/Matrix4.js";
-import Ray from "../../Core/Ray.js";
 import Resource from "../../Core/Resource.js";
 import RuntimeError from "../../Core/RuntimeError.js";
 import Pass from "../../Renderer/Pass.js";
@@ -45,8 +39,7 @@ import ModelUtility from "./ModelUtility.js";
 import oneTimeWarning from "../../Core/oneTimeWarning.js";
 import PntsLoader from "./PntsLoader.js";
 import StyleCommandsNeeded from "./StyleCommandsNeeded.js";
-import VertexAttributeSemantic from "../VertexAttributeSemantic.js";
-import Transforms from "../../Core/Transforms.js";
+import pickModel from "./pickModel.js";
 
 /**
  * <div class="notice">
@@ -160,6 +153,7 @@ import Transforms from "../../Core/Transforms.js";
  * @privateParam {boolean} [options.showCreditsOnScreen=false] Whether to display the credits of this model on screen.
  * @privateParam {SplitDirection} [options.splitDirection=SplitDirection.NONE] The {@link SplitDirection} split to apply to this model.
  * @privateParam {boolean} [options.projectTo2D=false] Whether to accurately project the model's positions in 2D. If this is true, the model will be projected accurately to 2D, but it will use more memory to do so. If this is false, the model will use less memory and will still render in 2D / CV mode, but its positions may be inaccurate. This disables minimumPixelSize and prevents future modification to the model matrix. This also cannot be set after the model has loaded.
+ * @privateParam {boolean} [options.enablePick=false] Whether to allow with CPU picking with <code>pick</code> when not using WebGL 2 or above. If using WebGL 2 or above, this option will be ignored. If using WebGL 1 and this is true, the <code>pick</code> operation will work correctly, but it will use more memory to do so. If running with WebGL 1 and this is false, the model will use less memory, but <code>pick</code> will always return <code>undefined</code>. This cannot be set after the model has loaded.
  * @privateParam {string|number} [options.featureIdLabel="featureId_0"] Label of the feature ID set to use for picking and styling. For EXT_mesh_features, this is the feature ID's label property, or "featureId_N" (where N is the index in the featureIds array) when not specified. EXT_feature_metadata did not have a label field, so such feature ID sets are always labeled "featureId_N" where N is the index in the list of all feature Ids, where feature ID attributes are listed before feature ID textures. If featureIdLabel is an integer N, it is converted to the string "featureId_N" automatically. If both per-primitive and per-instance feature IDs are present, the instance feature IDs take priority.
  * @privateParam {string|number} [options.instanceFeatureIdLabel="instanceFeatureId_0"] Label of the instance feature ID set used for picking and styling. If instanceFeatureIdLabel is set to an integer N, it is converted to the string "instanceFeatureId_N" automatically. If both per-primitive and per-instance feature IDs are present, the instance feature IDs take priority.
  * @privateParam {object} [options.pointCloudShading] Options for constructing a {@link PointCloudShading} object to control point attenuation based on geometric error and lighting.
@@ -463,6 +457,7 @@ function Model(options) {
 
   this._sceneMode = undefined;
   this._projectTo2D = defaultValue(options.projectTo2D, false);
+  this._enablePick = defaultValue(options.enablePick, false);
 
   this._skipLevelOfDetail = false;
   this._ignoreCommands = defaultValue(options.ignoreCommands, false);
@@ -2487,201 +2482,18 @@ Model.prototype.isClippingEnabled = function () {
   );
 };
 
-const scratchV0 = new Cartesian3();
-const scratchV1 = new Cartesian3();
-const scratchV2 = new Cartesian3();
-const scratchModelMatrix = new Matrix4();
-const scratchPickCartographic = new Cartographic();
-
 /**
  * Find an intersection between a ray and the model surface that was rendered. The ray must be given in world coordinates.
  *
  * @param {Ray} ray The ray to test for intersection.
  * @param {FrameState} frameState The frame state.
- * @param {boolean} [cullBackFaces=true] If false, back faces are not culled and will return an intersection if picked.
  * @param {Cartesian3|undefined} [result] The intersection or <code>undefined</code> if none was found.
  * @returns {Cartesian3|undefined} The intersection or <code>undefined</code> if none was found.
  *
  * @private
  */
-Model.prototype.pick = function (ray, frameState, cullBackFaces, result) {
-  if (frameState.mode === SceneMode.MORPHING) {
-    return;
-  }
-
-  let minT = Number.MAX_VALUE;
-  const modelMatrix = this.sceneGraph.computedModelMatrix;
-
-  // Check all the primitive positions
-  const nodes = this._sceneGraph._runtimeNodes;
-  for (let i = 0; i < nodes.length; i++) {
-    const node = nodes[i];
-    const instances = node.node.instances;
-    if (defined(instances)) {
-      // TODO: Instances
-      return;
-    }
-
-    const nodeComputedTransform = node.computedTransform;
-    let computedModelMatrix = Matrix4.multiplyTransformation(
-      modelMatrix,
-      nodeComputedTransform,
-      scratchModelMatrix
-    );
-    if (frameState.mode !== SceneMode.SCENE3D) {
-      computedModelMatrix = Transforms.basisTo2D(
-        frameState.mapProjection,
-        computedModelMatrix,
-        computedModelMatrix
-      );
-    }
-
-    for (let j = 0; j < node.node.primitives.length; j++) {
-      const primitive = node.node.primitives[j];
-      const positionAttribute = ModelUtility.getAttributeBySemantic(
-        primitive,
-        VertexAttributeSemantic.POSITION
-      );
-      const vertexCount = positionAttribute.count;
-
-      let indices = primitive.indices.typedArray;
-      if (!defined(indices)) {
-        const indicesBuffer = primitive.indices.buffer;
-        const indicesCount = primitive.indices.count;
-        if (defined(indicesBuffer) && frameState.context.webgl2) {
-          const useUint8Array = indicesBuffer.sizeInBytes === indicesCount;
-          indices = useUint8Array
-            ? new Uint8Array(indicesCount)
-            : IndexDatatype.createTypedArray(vertexCount, indicesCount);
-          indicesBuffer.getBufferData(indices, 0, 0, indicesCount);
-        }
-        primitive.indices.typedArray = indices;
-      }
-
-      let vertices = positionAttribute.typedArray;
-      let componentDatatype = positionAttribute.componentDatatype;
-      let attributeType = positionAttribute.type;
-
-      const quantization = positionAttribute.quantization;
-      if (defined(quantization)) {
-        componentDatatype = positionAttribute.quantization.componentDatatype;
-        attributeType = positionAttribute.quantization.type;
-      }
-
-      const numComponents = AttributeType.getNumberOfComponents(attributeType);
-      const elementCount = vertexCount * numComponents;
-
-      if (!defined(vertices)) {
-        const verticesBuffer = positionAttribute.buffer;
-
-        if (defined(verticesBuffer) && frameState.context.webgl2) {
-          vertices = ComponentDatatype.createTypedArray(
-            componentDatatype,
-            elementCount
-          );
-          verticesBuffer.getBufferData(
-            vertices,
-            positionAttribute.byteOffset,
-            0,
-            elementCount
-          );
-        }
-
-        if (quantization && positionAttribute.normalized) {
-          vertices = AttributeCompression.dequantize(
-            vertices,
-            componentDatatype,
-            attributeType,
-            vertexCount
-          );
-        }
-
-        positionAttribute.typedArray = vertices;
-      }
-
-      const indicesLength = indices.length;
-      for (let i = 0; i < indicesLength; i += 3) {
-        const i0 = indices[i];
-        const i1 = indices[i + 1];
-        const i2 = indices[i + 2];
-
-        const getPosition = (vertices, index, numComponents, result) => {
-          const i = index * numComponents;
-          result.x = vertices[i];
-          result.y = vertices[i + 1];
-          result.z = vertices[i + 2];
-
-          if (defined(quantization)) {
-            if (quantization.octEncoded) {
-              result = AttributeCompression.octDecodeInRange(
-                result,
-                quantization.normalizationRange,
-                result
-              );
-
-              if (quantization.octEncodedZXY) {
-                const x = result.x;
-                result.x = result.z;
-                result.z = result.y;
-                result.y = x;
-              }
-            } else {
-              result = Cartesian3.multiplyComponents(
-                result,
-                quantization.quantizedVolumeStepSize,
-                result
-              );
-
-              result = Cartesian3.add(
-                result,
-                quantization.quantizedVolumeOffset,
-                result
-              );
-            }
-          }
-
-          result = Matrix4.multiplyByPoint(computedModelMatrix, result, result);
-
-          return result;
-        };
-
-        const v0 = getPosition(vertices, i0, numComponents, scratchV0);
-        const v1 = getPosition(vertices, i1, numComponents, scratchV1);
-        const v2 = getPosition(vertices, i2, numComponents, scratchV2);
-
-        const t = IntersectionTests.rayTriangleParametric(
-          ray,
-          v0,
-          v1,
-          v2,
-          cullBackFaces
-        );
-
-        if (defined(t)) {
-          if (t < minT && t >= 0.0) {
-            minT = t;
-          }
-        }
-      }
-    }
-  }
-
-  if (minT === Number.MAX_VALUE) {
-    return undefined;
-  }
-
-  result = Ray.getPoint(ray, minT, result);
-  if (frameState.mode !== SceneMode.SCENE3D) {
-    Cartesian3.fromElements(result.y, result.z, result.x, result);
-
-    const projection = frameState.mapProjection;
-    const ellipsoid = projection.ellipsoid;
-
-    const cart = projection.unproject(result, scratchPickCartographic);
-    ellipsoid.cartographicToCartesian(cart, result);
-  }
-
-  return result;
+Model.prototype.pick = function (ray, frameState, result) {
+  return pickModel(this, ray, frameState, result);
 };
 
 /**
@@ -2843,6 +2655,7 @@ Model.prototype.destroyModelResources = function () {
  * @param {boolean} [options.showCreditsOnScreen=false] Whether to display the credits of this model on screen.
  * @param {SplitDirection} [options.splitDirection=SplitDirection.NONE] The {@link SplitDirection} split to apply to this model.
  * @param {boolean} [options.projectTo2D=false] Whether to accurately project the model's positions in 2D. If this is true, the model will be projected accurately to 2D, but it will use more memory to do so. If this is false, the model will use less memory and will still render in 2D / CV mode, but its positions may be inaccurate. This disables minimumPixelSize and prevents future modification to the model matrix. This also cannot be set after the model has loaded.
+ * @param {boolean} [options.enablePick=false] Whether to allow with CPU picking with <code>pick</code> when not using WebGL 2 or above. If using WebGL 2 or above, this option will be ignored. If using WebGL 1 and this is true, the <code>pick</code> operation will work correctly, but it will use more memory to do so. If running with WebGL 1 and this is false, the model will use less memory, but <code>pick</code> will always return <code>undefined</code>. This cannot be set after the model has loaded.
  * @param {string|number} [options.featureIdLabel="featureId_0"] Label of the feature ID set to use for picking and styling. For EXT_mesh_features, this is the feature ID's label property, or "featureId_N" (where N is the index in the featureIds array) when not specified. EXT_feature_metadata did not have a label field, so such feature ID sets are always labeled "featureId_N" where N is the index in the list of all feature Ids, where feature ID attributes are listed before feature ID textures. If featureIdLabel is an integer N, it is converted to the string "featureId_N" automatically. If both per-primitive and per-instance feature IDs are present, the instance feature IDs take priority.
  * @param {string|number} [options.instanceFeatureIdLabel="instanceFeatureId_0"] Label of the instance feature ID set used for picking and styling. If instanceFeatureIdLabel is set to an integer N, it is converted to the string "instanceFeatureId_N" automatically. If both per-primitive and per-instance feature IDs are present, the instance feature IDs take priority.
  * @param {object} [options.pointCloudShading] Options for constructing a {@link PointCloudShading} object to control point attenuation and lighting.
@@ -2937,6 +2750,7 @@ Model.fromGltfAsync = async function (options) {
     upAxis: options.upAxis,
     forwardAxis: options.forwardAxis,
     loadAttributesFor2D: options.projectTo2D,
+    enablePick: options.enablePick,
     loadIndicesForWireframe: options.enableDebugWireframe,
     loadPrimitiveOutline: options.enableShowOutline,
     loadForClassification: defined(options.classificationType),
@@ -3013,6 +2827,7 @@ Model.fromB3dm = async function (options) {
     upAxis: options.upAxis,
     forwardAxis: options.forwardAxis,
     loadAttributesFor2D: options.projectTo2D,
+    enablePick: options.enablePick,
     loadIndicesForWireframe: options.enableDebugWireframe,
     loadPrimitiveOutline: options.enableShowOutline,
     loadForClassification: defined(options.classificationType),
@@ -3068,6 +2883,7 @@ Model.fromI3dm = async function (options) {
     upAxis: options.upAxis,
     forwardAxis: options.forwardAxis,
     loadAttributesFor2D: options.projectTo2D,
+    enablePick: options.enablePick,
     loadIndicesForWireframe: options.enableDebugWireframe,
     loadPrimitiveOutline: options.enableShowOutline,
   };
@@ -3201,6 +3017,7 @@ function makeModelOptions(loader, modelType, options) {
     showCreditsOnScreen: options.showCreditsOnScreen,
     splitDirection: options.splitDirection,
     projectTo2D: options.projectTo2D,
+    enablePick: options.enablePick,
     featureIdLabel: options.featureIdLabel,
     instanceFeatureIdLabel: options.instanceFeatureIdLabel,
     pointCloudShading: options.pointCloudShading,