+++ packages/engine/Specs/Scene/Cesium3DTilesetSpec.js
@@ -46,6 +46,7 @@ import Cesium3DTilesTester from "../../../../Specs/Cesium3DTilesTester.js";
 import createScene from "../../../../Specs/createScene.js";
 import generateJsonBuffer from "../../../../Specs/generateJsonBuffer.js";
 import pollToPromise from "../../../../Specs/pollToPromise.js";
+import Ellipsoid from "../../Source/Core/Ellipsoid.js";
 
 describe(
   "Scene/Cesium3DTileset",
@@ -2412,6 +2413,183 @@ describe(
           expect(arg.url).toContain("parent.b3dm");
           expect(arg.message).toBeDefined();
         });
+    });
+
+    it("picks", async function () {
+      const tileset = await Cesium3DTilesTester.loadTileset(scene, tilesetUrl, {
+        enablePick: !scene.frameState.context.webgl2,
+      });
+      viewRootOnly();
+      scene.renderForSpecs();
+
+      const ray = scene.camera.getPickRay(
+        new Cartesian2(
+          scene.drawingBufferWidth / 2.0,
+          scene.drawingBufferHeight / 2.0
+        )
+      );
+
+      const expected = new Cartesian3(
+        1215026.8094312553,
+        -4736367.339076743,
+        4081652.238842398
+      );
+      expect(tileset.pick(ray, scene.frameState)).toEqualEpsilon(
+        expected,
+        CesiumMath.EPSILON12
+      );
+    });
+
+    it("picks tileset of tilesets", async function () {
+      const tileset = await Cesium3DTilesTester.loadTileset(
+        scene,
+        tilesetOfTilesetsUrl,
+        {
+          enablePick: !scene.frameState.context.webgl2,
+        }
+      );
+      viewRootOnly();
+      scene.renderForSpecs();
+
+      const ray = scene.camera.getPickRay(
+        new Cartesian2(
+          scene.drawingBufferWidth / 2.0,
+          scene.drawingBufferHeight / 2.0
+        )
+      );
+
+      const expected = new Cartesian3(
+        1215026.8094312553,
+        -4736367.339076743,
+        4081652.238842398
+      );
+      expect(tileset.pick(ray, scene.frameState)).toEqualEpsilon(
+        expected,
+        CesiumMath.EPSILON12
+      );
+    });
+
+    it("picks instanced tileset", async function () {
+      const tileset = await Cesium3DTilesTester.loadTileset(
+        scene,
+        instancedUrl,
+        {
+          enablePick: !scene.frameState.context.webgl2,
+        }
+      );
+      viewInstances();
+      scene.renderForSpecs();
+
+      const ray = scene.camera.getPickRay(
+        new Cartesian2(
+          scene.drawingBufferWidth / 2.0,
+          scene.drawingBufferHeight / 2.0
+        )
+      );
+
+      const expected = new Cartesian3(
+        1215015.7820120894,
+        -4736324.352446682,
+        4081615.004915994
+      );
+      expect(tileset.pick(ray, scene.frameState)).toEqualEpsilon(
+        expected,
+        CesiumMath.EPSILON12
+      );
+    });
+
+    it("picks translucent tileset", async function () {
+      const tileset = await Cesium3DTilesTester.loadTileset(
+        scene,
+        translucentUrl,
+        {
+          enablePick: !scene.frameState.context.webgl2,
+        }
+      );
+      viewAllTiles();
+      scene.renderForSpecs();
+
+      const ray = scene.camera.getPickRay(
+        new Cartesian2(
+          scene.drawingBufferWidth / 2.0,
+          scene.drawingBufferHeight / 2.0
+        )
+      );
+
+      const expected = new Cartesian3(
+        1215013.1035421563,
+        -4736313.911345786,
+        4081605.96109977
+      );
+      expect(tileset.pick(ray, scene.frameState)).toEqualEpsilon(
+        expected,
+        CesiumMath.EPSILON12
+      );
+    });
+
+    it("picks tileset with transforms", async function () {
+      const tileset = await Cesium3DTilesTester.loadTileset(
+        scene,
+        tilesetWithTransformsUrl,
+        {
+          enablePick: !scene.frameState.context.webgl2,
+        }
+      );
+      viewAllTiles();
+      scene.renderForSpecs();
+
+      const ray = scene.camera.getPickRay(
+        new Cartesian2(
+          scene.drawingBufferWidth / 2.0,
+          scene.drawingBufferHeight / 2.0
+        )
+      );
+
+      const expected = new Cartesian3(
+        1215013.8353220497,
+        -4736316.763939952,
+        4081608.4319443353
+      );
+      expect(tileset.pick(ray, scene.frameState)).toEqualEpsilon(
+        expected,
+        CesiumMath.EPSILON12
+      );
+    });
+
+    it("picking point cloud tileset returns undefined", async function () {
+      const tileset = await Cesium3DTilesTester.loadTileset(
+        scene,
+        pointCloudUrl,
+        {
+          enablePick: !scene.frameState.context.webgl2,
+        }
+      );
+      viewAllTiles();
+      scene.renderForSpecs();
+
+      const ray = scene.camera.getPickRay(
+        new Cartesian2(
+          scene.drawingBufferWidth / 2.0,
+          scene.drawingBufferHeight / 2.0
+        )
+      );
+
+      expect(tileset.pick(ray, scene.frameState)).toBeUndefined();
+    });
+
+    it("getHeight samples height at a cartographic position", async function () {
+      const tileset = await Cesium3DTilesTester.loadTileset(scene, tilesetUrl, {
+        enablePick: !scene.frameState.context.webgl2,
+      });
+      viewRootOnly();
+      await Cesium3DTilesTester.waitForTilesLoaded(scene, tileset);
+      scene.renderForSpecs();
+
+      const center = Ellipsoid.WGS84.cartesianToCartographic(
+        tileset.boundingSphere.center
+      );
+      const height = tileset.getHeight(center, scene);
+      expect(height).toEqualEpsilon(78.1558019795064, CesiumMath.EPSILON12);
     });
 
     it("destroys", function () {